requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
blis==0.7.11
boto3==1.28.39
botocore==1.31.39
catalogue==2.0.10
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
cloudpathlib==0.15.1
colorama==0.4.6
comtypes==1.2.0
confection==0.1.3
cymem==2.0.8
dataclasses-json==0.6.2
Django==4.2.4
django-cors-headers==4.2.0
django-environ==0.11.1
django-rest-passwordreset==1.3.0
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-api-tracking==1.8.4
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl#sha256=6215d71a3212690e9aec49408a27e3fe6ad7cd6c715476e93d70dc784041e93e
frozenlist==1.4.0
greenlet==3.0.1
idna==3.4
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.340
langcodes==3.3.0
langsmith==0.0.66
MarkupSafe==2.1.3
marshmallow==3.20.1
multidict==6.0.4
murmurhash==1.0.10
mypy-extensions==1.0.0
nltk==3.8.1
numpy==1.26.0
openai==0.27.10
packaging==23.1
pathy==0.10.2
Pillow==10.0.0
preshed==3.0.9
psycopg2-binary==2.9.7
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.8.0
pypiwin32==223
python-dateutil==2.8.2
python-decouple==3.8
python-environ==0.4.54
pyttsx3==2.90
pytz==2023.3
pywin32==305
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
s3transfer==0.6.2
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
spacy==3.7.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.10.0
SQLAlchemy==2.0.23
sqlparse==0.4.4
srsly==2.4.8
tenacity==8.2.3
thinc==8.2.1
tqdm==4.66.1
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.16
wasabi==1.1.2
Wave==0.0.2
weasel==0.3.2
yarl==1.9.2
